Remove unreasonable unit tests
